backend/requirements.txt
# AWS SDK
boto3==1.34.0
aioboto3==12.3.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.1.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.0
moto==4.2.11
httpx==0.25.2

# Local Development API (FastAPI)
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
